// src/app/modules/order-dispatch/Models/all-inv-cnt.model.ts
export class AllInvCnt {
    TodaysWithOldOpen : number =0;
    TotalInvoices: number = 0;
    CancelInvCtn: number = 0;
    PendingInvCtn: number = 0;
    OnPriorityCtn: number = 0;
    PackerConcern: number = 0;
    GatpassGenCtn: number = 0;
    PendingLR : number = 0;
    IsStockTransferCtn: number = 0;
    StkPrint: number = 0;
    LocalMode: number = 0;
    OtherCity: number = 0;
    ByHand: number = 0;
}